use-cases/utilities/data-loader-lambda/requirements.txt
boto3>=1.26.0
pymysql>=1.0.2
psycopg2-binary>=2.9.5
